More testing and better testing for sessions, better semantics and file structures as in a single mutations file
