+++ authn/webapi/sessions/hooks/mutations/mutations.go
@@ -1,19 +1,16 @@
 package mutations
 
 import (
+	"encoding/json"
 	err "errors"
+	"net/http"
 
-	"webapi/sessions/hooks/requests"
 	"webapi/interfaces/jwtx"
+	"webapi/sessions/hooks/constants"
+	"webapi/sessions/hooks/errors"
+	"webapi/sessions/hooks/requests"
+	"webapi/sessions/hooks/responses"
 	"webapi/sessions/sessionsx"
-	"webapi/sessions/constants"
-)
-
-var (
-	CreateGuestSessionErrorMessage = "error creating guest session"
-	InvalidRequestProvided         = "invalid request provided"
-	InvalidSessionProvided         = "invalid session provided"
-	UnableToMarshalSession         = "unable to marshal session"
 )
 
 func updateGenericSession(p *requests.Update) (*sessionsx.Session, error) {
@@ -35,4 +32,431 @@ func updateGenericSession(p *requests.Update) (*sessionsx.Session, error) {
 	}
 
 	return session, nil
+}
+
+func CreateGuestSession(w http.ResponseWriter, requestBody *requests.Body) {	
+	if requestBody == nil || requestBody.Params == nil {
+		errors.BadRequest(w, &responses.Errors{
+			Body: &errors.BadRequestFail,
+		})
+		return
+	}
+
+	bytes, _ := json.Marshal(requestBody.Params)
+	var params requests.SessionParams
+	errParamsMarshal := json.Unmarshal(bytes, &params)
+	if errParamsMarshal != nil {
+		errAsStr := errParamsMarshal.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Body: &errors.BadRequestFail,
+			Default: &errAsStr,
+		})
+		return
+	}
+
+	session, errSession := sessionsx.Create(&sessionsx.CreateParams{
+		Environment: params.Environment,
+		Claims: *sessionsx.CreateGuestSessionClaims(),
+	})
+
+	if errSession == nil {
+		payload := responses.Session{
+			SessionToken: session.SessionToken,
+		}
+		body := responses.Body{
+			Session: &payload,
+		}
+
+		w.Header().Set("Content-Type", "application/json")
+		json.NewEncoder(w).Encode(&body)
+		return
+	}
+
+	errorAsStr := errSession.Error()
+	errors.BadRequest(w, &responses.Errors{
+		Session: &errors.CreateGuestSessionErrorMessage,
+		Default: &errorAsStr,
+	})
+}
+
+func CreateDocumentSession(w http.ResponseWriter, requestBody *requests.Body) {
+	if requestBody == nil || requestBody.Params == nil {
+		errors.BadRequest(w, &responses.Errors{
+			Body: &errors.BadRequestFail,
+		})
+		return
+	}
+
+	bytes, _ := json.Marshal(requestBody.Params)
+	var params requests.SessionParams
+	errParamsMarshal := json.Unmarshal(bytes, &params)
+	if errParamsMarshal != nil {
+		errAsStr := errParamsMarshal.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Body: &errors.BadRequestFail,
+			Default: &errAsStr,
+		})
+		return
+	}
+
+	session, errSession := sessionsx.Create(&sessionsx.CreateParams{
+		Environment: params.Environment,
+		Claims: *sessionsx.CreateDocumentSessionClaims(),
+	})
+	if errSession == nil {
+		payload := responses.Session{
+			SessionToken: session.SessionToken,
+		}
+		body := responses.Body{
+			Session: &payload,
+		}
+
+		w.Header().Set("Content-Type", "application/json")
+		json.NewEncoder(w).Encode(&body)
+		return
+	}
+
+	errorAsStr := errSession.Error()
+	errors.BadRequest(w, &responses.Errors{
+		Session: &errors.CreateGuestSessionErrorMessage,
+		Default: &errorAsStr,
+	})
+}
+
+
+func CreateCreateAccountSession(w http.ResponseWriter, requestBody *requests.Body) {
+	if requestBody == nil || requestBody.Params == nil {
+		errors.BadRequest(w, &responses.Errors{
+			Session: &errors.InvalidSessionCredentials,
+			Body: &errors.BadRequestFail,
+		})
+		return
+	}
+
+	bytes, _ := json.Marshal(requestBody.Params)
+	var params requests.AccountParams
+	errParamsMarshal := json.Unmarshal(bytes, &params)
+	if errParamsMarshal != nil {
+		errAsStr := errParamsMarshal.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Body: &errors.BadRequestFail,
+			Default: &errAsStr,
+		})
+		return
+	}
+
+	sessionParams, errSessionParams := sessionsx.CreateAccountCreationSessionClaims(
+		&params,
+	)
+
+	if errSessionParams != nil {
+		errors.CustomErrorResponse(w, errors.InvalidSessionCredentials)
+		return
+	}
+
+	session, errSession := sessionsx.Create(&sessionsx.CreateParams{
+		Environment: params.Environment,
+		Claims: *sessionParams,
+	})
+
+	if errSession == nil {
+		marshalledJSON, errMarshal := json.Marshal(&responses.Session{
+			SessionToken: session.SessionToken,
+		})
+		if errMarshal == nil {
+			w.Header().Set("Content-Type", "application/json")
+			json.NewEncoder(w).Encode(&marshalledJSON)
+			return
+		}
+
+		errors.CustomErrorResponse(w, errors.UnableToMarshalSession)
+		return
+	}
+
+	errorAsStr := errSession.Error()
+	errors.BadRequest(w, &responses.Errors{
+		Session: &errors.CreateGuestSessionErrorMessage,
+		Default: &errorAsStr,
+	})
+}
+
+func CreateUpdateEmailSession(w http.ResponseWriter, requestBody *requests.Body) {
+	if requestBody == nil || requestBody.Params == nil {
+		errors.BadRequest(w, &responses.Errors{
+			Session: &errors.InvalidSessionCredentials,
+			Body: &errors.BadRequestFail,
+		})
+		return
+	}
+
+	bytes, _ := json.Marshal(requestBody.Params)
+	var params requests.AccountParams
+	errParamsMarshal := json.Unmarshal(bytes, &params)
+	if errParamsMarshal != nil {
+		errAsStr := errParamsMarshal.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Body: &errors.BadRequestFail,
+			Default: &errAsStr,
+		})
+		return
+	}
+
+	userSessionToken, errUserSessionToken := sessionsx.CreateUpdatePasswordSessionClaims(
+		&params,
+	)
+	if errUserSessionToken != nil {
+		errorAsStr := errUserSessionToken.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Session: &errors.InvalidSessionCredentials,
+			Default: &errorAsStr,
+		})
+		return
+	}
+
+	session, errSession := sessionsx.Create(&sessionsx.CreateParams{
+		Environment: params.Environment,
+		Claims: *userSessionToken,
+	})
+
+	if errSession == nil {
+		marshalledJSON, errMarshal := json.Marshal(
+			&responses.Session{
+				SessionToken: session.SessionToken,
+			},
+		)
+		if errMarshal == nil {
+			w.Header().Set("Content-Type", "application/json")
+			json.NewEncoder(w).Encode(&marshalledJSON)
+			return
+		}
+		
+		errors.CustomErrorResponse(w, errors.UnableToMarshalSession)
+		return
+	}
+
+	errorAsStr := errSession.Error()
+	errors.BadRequest(w, &responses.Errors{
+		Session: &errors.CreateGuestSessionErrorMessage,
+		Default: &errorAsStr,
+	})
+}
+
+func CreateUpdatePasswordSession(w http.ResponseWriter, requestBody *requests.Body) {
+	if requestBody == nil || requestBody.Params == nil {
+		errors.BadRequest(w, &responses.Errors{
+			Session: &errors.InvalidSessionCredentials,
+			Body: &errors.BadRequestFail,
+		})
+		return
+	}
+
+	bytes, _ := json.Marshal(requestBody.Params)
+	var params requests.AccountParams
+	errParamsMarshal := json.Unmarshal(bytes, &params)
+	if errParamsMarshal != nil {
+		errAsStr := errParamsMarshal.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Body: &errors.BadRequestFail,
+			Default: &errAsStr,
+		})
+		return
+	}
+
+	userSessionToken, errUserSessionToken := sessionsx.CreateUpdatePasswordSessionClaims(
+		&params,
+	)
+	if errUserSessionToken != nil {
+		errorAsStr := errUserSessionToken.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Session: &errors.InvalidSessionCredentials,
+			Default: &errorAsStr,
+		})
+		return
+	}
+
+	session, errSession := sessionsx.Create(&sessionsx.CreateParams{
+		Environment: params.Environment,
+		Claims: *userSessionToken,
+	})
+
+	if errSession == nil {
+		marshalledJSON, errMarshal := json.Marshal(&responses.Session{
+			SessionToken: session.SessionToken,
+		})
+		if errMarshal == nil {
+			w.Header().Set("Content-Type", "application/json")
+			json.NewEncoder(w).Encode(&marshalledJSON)
+			return
+		}
+
+		errors.CustomErrorResponse(w, errors.UnableToMarshalSession)
+		return
+	}
+
+	errorAsStr := errSession.Error()
+	errors.BadRequest(w, &responses.Errors{
+		Session: &errors.CreateGuestSessionErrorMessage,
+		Default: &errorAsStr,
+	})
+}
+
+
+func CreatePublicSession(w http.ResponseWriter, requestBody *requests.Body) {
+	if requestBody == nil || requestBody.Params == nil {
+		errors.BadRequest(w, &responses.Errors{
+			Session: &errors.InvalidSessionCredentials,
+			Body: &errors.BadRequestFail,
+		})
+		return
+	}
+
+	bytes, _ := json.Marshal(requestBody.Params)
+	var params requests.UserParams
+	errParamsMarshal := json.Unmarshal(bytes, &params)
+	if errParamsMarshal != nil {
+		errAsStr := errParamsMarshal.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Body: &errors.BadRequestFail,
+			Default: &errAsStr,
+		})
+		return
+	}
+
+	userSessionToken, errUserSessionToken := sessionsx.CreateUserSessionClaims(
+		&params,
+	)
+	if errUserSessionToken != nil {
+		errorAsStr := errUserSessionToken.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Session: &errors.InvalidSessionCredentials,
+			Default: &errorAsStr,
+		})
+		return
+	}
+
+	userSession, errUserSession := sessionsx.Create(&sessionsx.CreateParams{
+		Environment: params.Environment,
+		Claims:	*userSessionToken,
+	})
+
+	if errUserSession == nil {
+		marshalledJSON, errMarshal := json.Marshal(
+			&responses.Session{
+				SessionToken: userSession.SessionToken,
+			},
+		)
+		if errMarshal == nil {
+			w.Header().Set("Content-Type", "application/json")
+			json.NewEncoder(w).Encode(&marshalledJSON)
+			return
+		}
+
+		errors.CustomErrorResponse(w, errors.UnableToMarshalSession)
+		return
+	}
+
+	errorAsStr := errUserSession.Error()
+	errors.BadRequest(w, &responses.Errors{
+		Session: &errors.UnableToCreatePublicSession,
+		Default: &errorAsStr,
+	})
+}
+
+
+func UpdateSession(w http.ResponseWriter, requestBody *requests.Body) {
+	if requestBody == nil || requestBody.Params == nil {
+		errors.BadRequest(w, &responses.Errors{
+			Session: &errors.UnableToUpdateSession,
+			Body: &errors.BadRequestFail,
+		})
+		return
+	}
+
+	bytes, _ := json.Marshal(requestBody.Params)
+	var params requests.Update
+	errParamsMarshal := json.Unmarshal(bytes, &params)
+	if errParamsMarshal != nil {
+		errAsStr := errParamsMarshal.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Body: &errors.BadRequestFail,
+			Default: &errAsStr,
+		})
+		return
+	}
+
+	session, errSession := updateGenericSession(&params)
+	if errSession != nil {
+		errAsStr := errSession.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Session: &errors.InvalidSessionProvided,
+			Default: &errAsStr,
+		})
+		return
+	}
+
+	if errSession == nil {
+		marshalledJSON, errMarshal := json.Marshal(
+			&responses.Session{
+				SessionToken: session.SessionToken,
+			},
+		)
+		if errMarshal != nil {
+			errAsStr := errMarshal.Error()
+			errors.BadRequest(w, &responses.Errors{
+				Session: &errors.UnableToMarshalSession,
+				Default: &errAsStr,
+			})
+			return
+		}
+
+		w.Header().Set("Content-Type", "application/json")
+		json.NewEncoder(w).Encode(&marshalledJSON)
+		return
+	}
+
+	errAsStr := errSession.Error()
+	errors.BadRequest(w, &responses.Errors{
+		Session: &errors.UnableToUpdateSession,
+		Default: &errAsStr,
+	})
+}
+
+func DeleteSession(w http.ResponseWriter, requestBody *requests.Body) {
+	if requestBody == nil || requestBody.Params == nil {
+		errors.BadRequest(w, &responses.Errors{
+			Session: &errors.InvalidSessionCredentials,
+			Body: &errors.BadRequestFail,
+		})
+		return
+	}
+
+	bytes, _ := json.Marshal(requestBody.Params)
+	var params requests.Delete
+	errParamsMarshal := json.Unmarshal(bytes, &params)
+	if errParamsMarshal != nil {
+		errAsStr := errParamsMarshal.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Body: &errors.BadRequestFail,
+			Default: &errAsStr,
+		})
+		return
+	}
+
+	result, errResponseBody := sessionsx.Delete(&params)
+	if errResponseBody != nil {
+		errAsStr := errResponseBody.Error()
+		errors.BadRequest(w, &responses.Errors{
+			Session: &errors.InvalidSessionProvided,
+			Default: &errAsStr,
+		})
+		return
+	}
+
+	if result == true {
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusOK)
+		return
+	}
+
+	errors.CustomErrorResponse(w, errors.InvalidSessionProvided)
 }
